# Divrem speedup and tests