--- a/base/div.jl
+++ b/base/div.jl
@@ -134,20 +134,7 @@ julia> divrem(7,3)
 """
 divrem(x, y) = divrem(x, y, RoundToZero)
 
-#avoids calling rem for Integers-Integers,
-#a-d*b not precise for Floats - AbstractFloat, AbstractIrrational. Rationals are still slower
-function divrem(a::Integer, b::Integer, r::RoundingMode)
-    if r === RoundToZero
-        d = div(a, b)
-        (d, a - d*b)
-    elseif r === RoundDown
-        d = fld(a, b)
-        (d, a - d*b)
-    else
-        d = div(a, b, r)
-        (d, a - d*b)
-    end
-end
+
 function divrem(a, b, r::RoundingMode)
     if r === RoundToZero
         # For compat. Remove in 2.0.
@@ -157,6 +144,25 @@ function divrem(a, b, r::RoundingMode)
         (fld(a, b),mod(a,b))
     else
         (div(a,b,r), rem(a,b,r))
+    end
+end
+#avoids calling rem for Integers-Integers (all modes),
+#a-d*b not precise for Floats - AbstractFloat, AbstractIrrational. Rationals are still slower
+function divrem(a::Integer, b::Integer, r::Union{typeof(RoundUp),
+                                                typeof(RoundDown),
+                                                typeof(RoundToZero)})
+    if r === RoundToZero
+        # For compat. Remove in 2.0.
+        d = div(a, b)
+        (d, a - d*b)
+    elseif r === RoundDown
+        # For compat. Remove in 2.0.
+        d = fld(a, b)
+        (d, a - d*b)
+    elseif r == RoundUp
+        # For compat. Remove in 2.0.
+        d = div(a, b, r)
+        (d, a - d*b)
     end
 end
 function divrem(x::Integer, y::Integer, rnd::typeof(RoundNearest))