--- a/test/int.jl
+++ b/test/int.jl
@@ -339,8 +339,8 @@ let i = MyInt26779(1)
 end
 
 @testset "rounding division" begin
-    for x = -50:50
-        for y = 1:50
+    for x = -100:100
+        for y = 1:100
             for rnd in (RoundNearest, RoundNearestTiesAway, RoundNearestTiesUp)
                 @test div(x,y,rnd) == round(x/y,rnd)
                 @test div(x,-y,rnd) == round(x/-y,rnd)