[#22219](https://github.com/JuliaLang/julia/issues/22219#issue-233492600)
- Replaced rem call with `a-d*b`. Reduced div calls from 2-->1 (code in: `base\div.jl`)
Int-Int : Works for all rounding modes and speeds up by a factor of 1.6x
Floats - Abstract Floats, Abstract Irrational : Works but there is marginal error in Float64, BigFloat. (Not added)
Rationals - `a-d*b` is more expensive than normal div and rem calls. (Not added)

- Also added tests for divrem in rounding modes in `test\numbers.jl`

Passes all the tests for div, divrem, rem ` SUCCESS`
Also speeds divrem by a factor of 2
```
function divrem!(a, b, c, d)
    for i = 1:100_000
        c[i], d[i] = divrem(a[i], b[i])
    end
end

function divrem2!(a, b, c, d)
    for i = 1:100_000
        c[i] = div(a[i], b[i])
        d[i] = rem(a[i], b[i])
    end
end

function divrem3!(a, b, c, d)
    for i = 1:100_000
        c[i] = div(a[i], b[i])
        d[i] = a[i] - b[i] * c[i]
    end
end

a, b = rand(0:500_000, 100_000), rand(1:250_000, 100_000);
c, d = zeros(Int, 100_000), zeros(Int, 100_000)


julia> @benchmark divrem!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     834.000 μs (0.00% GC)
  median time:      837.300 μs (0.00% GC)
  mean time:        851.657 μs (0.00% GC)
  maximum time:     1.048 ms (0.00% GC)
  --------------
  samples:          100
  evals/sample:     1

julia> @benchmark divrem2!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.479 ms (0.00% GC)
  median time:      1.480 ms (0.00% GC)
  mean time:        1.499 ms (0.00% GC)
  maximum time:     1.716 ms (0.00% GC)
  --------------
  samples:          100
  evals/sample:     1

julia> @benchmark divrem3!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     839.100 μs (0.00% GC)
  median time:      843.400 μs (0.00% GC)
  mean time:        856.651 μs (0.00% GC)
  maximum time:     1.047 ms (0.00% GC)
  --------------
  samples:          100
  evals/sample:     1
```