@oscardssmith 2020-12-30
Have you tested speed for `BigInt`? I assume it will still be faster, but it often can cause corner cases. Other than that, this looks great!