@musm 2020-12-30
> Added misc changes like: Removed useless spaces in base\div.jl

These changes are orthogonal to the PR.