@r7rohan 2020-12-30
There was a trivial typo error fixed it.

Tested it for random BigInt, they show the same trend.
```
a, b = rand(0:500000000000000000000000000000000000000000000000000000000000000000000000000, 100000), rand(1:500000000000000000000000000000000000000000000000000000000000000000000000000, 100000);
c, d = zeros(BigInt, 100_000), zeros(BigInt, 100_000);

julia> @benchmark divrem!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  9.92 MiB
  allocs estimate:  500000
  --------------
  minimum time:     28.940 ms (0.00% GC)
  median time:      31.278 ms (0.00% GC)
  mean time:        46.815 ms (18.10% GC)
  maximum time:     435.204 ms (51.40% GC)
  --------------
  samples:          100
  evals/sample:     1

julia> @benchmark divrem2!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  9.92 MiB
  allocs estimate:  500000
  --------------
  minimum time:     32.511 ms (0.00% GC)
  median time:      36.163 ms (0.00% GC)
  mean time:        50.150 ms (16.95% GC)
  maximum time:     418.931 ms (53.83% GC)
  --------------
  samples:          100
  evals/sample:     1

julia> @benchmark divrem3!($a,$b,$c,$d) samples=100
BenchmarkTools.Trial:
  memory estimate:  16.02 MiB
  allocs estimate:  649972
  --------------
  minimum time:     31.629 ms (0.00% GC)
  median time:      33.585 ms (0.00% GC)
  mean time:        61.267 ms (24.99% GC)
  maximum time:     448.482 ms (51.84% GC)
  --------------
  samples:          100
  evals/sample:     1
```