@Keno 2021-01-08
This could use a PkgEval before merging. I know it doesn't seem like it, but the fallback chains for `div`-adjacent methods are very fragile because we allow packages to intercept both the 2 and 3-argument variants and want to provide various fallback. This needs to be cleaned up in 2.0, but not much we can do until then.