@r7rohan 2021-02-01
Sorry for late reply, I don't know much about PkgEval...so would appreciate help in that part. Apart from that, it seems to work well for divrem int-int types