@nanosoldier 2021-03-07
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/39045#issuecomment-792378088):
```
NanosoldierError: failed to run tests: Build for julia on x86_64-linux-gnu did not complete successfully

```
Logs and partial data can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/pkgeval/by_hash/24d688a_vs_5fab42a/report.md)
cc @maleadt