@nanosoldier 2021-03-08
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/39045#issuecomment-792908248) has completed - possible new issues were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/pkgeval/by_hash/24d688a_vs_86387a8/report.md). cc @maleadt