@KristofferC 2021-03-09
@nanosoldier `runtests(["AlphaStableDistributions", "BoltzmannMachines", "FilesystemDatastructures", "GPLinearODEMaker", "GeoDataFrames", "JetPackWaveFD", "Jusdl", "KissMCMC", "RiskAdjustedLinearizations", "Tar", "TensorToolbox"], vs = ":master")`