@vtjnash 2021-04-19
Seems to be working, but I'm confused by the line comments: why is all of the new code deprecated and marked for deletion in v2?