@vtjnash 2021-04-22
It seems to be working, I'm just confused by the content of the comments in the code, and was hoping you could clarify what they meant (I can make minor changes, if that's what's needed).