@Keno 2021-08-17
In v2, all primitive integer types should implement three argument div(rem) and the two argument versions should just be fallbacks to those. At the moment, we have various complicated messes to handle situations where people overload the various two argument methods.